+++ lambdas/reconciliation-report-migration/src/index.ts
@@ -1,7 +1,7 @@
 import { getKnexClient } from '@cumulus/db';
 import Logger from '@cumulus/logger';
 
-import { migrateAsyncOperations } from './async-operations';
+import { migrateReconciliationReports } from './reconciliation-reports';
 import { MigrationSummary } from './types';
 
 const logger = new Logger({ sender: '@cumulus/data-migration1' });
@@ -14,7 +14,7 @@ export const handler = async (event: HandlerEvent): Promise<MigrationSummary> =>
   const knex = await getKnexClient({ env });
 
   try {
-    const migrationSummary = await migrateAsyncOperations(env, knex);
+    const migrationSummary = await migrateReconciliationReports(env, knex);
     logger.info(JSON.stringify(migrationSummary));
     return migrationSummary;
   } finally {